Update the network API commucation.
