+++ app/src/main/java/ru/sandbox/androidacademyapp/api/MoviesApi.kt
@@ -1,0 +1,58 @@
+package ru.sandbox.androidacademyapp.api
+
+import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.json.Json
+import okhttp3.MediaType.Companion.toMediaType
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import retrofit2.http.GET
+import retrofit2.http.Path
+import ru.sandbox.androidacademyapp.BuildConfig
+import ru.sandbox.androidacademyapp.data.Actor
+import ru.sandbox.androidacademyapp.data.Movie
+
+interface MoviesApi {
+
+    @GET("movie/popular?page=1")
+    suspend fun getMovies(): MovieListResponse
+
+    @GET("movie/{movie_id}")
+    suspend fun getMovieDetails(@Path("movie_id") movie_id: Int): Movie
+
+    @GET("movie/{movie_id}/credits")
+    suspend fun getMovieActors(@Path("movie_id") movie_id: Int): ActorListResponse
+
+    @Serializable
+    data class ActorListResponse(
+        @SerialName("cast")
+        val actors: List<Actor>
+    )
+
+    @Serializable
+    data class MovieListResponse(
+        @SerialName("results")
+        val movies: List<Movie>
+    )
+
+    companion object {
+        private val json = Json { ignoreUnknownKeys = true }
+
+        @Suppress("EXPERIMENTAL_API_USAGE")
+        fun create(): MoviesApi {
+            val client = OkHttpClient().newBuilder()
+                .addInterceptor(HttpLoggingInterceptor().setLevel(HttpLoggingInterceptor.Level.BODY))
+                .addInterceptor(MoviesApiHeaderInterceptor())
+                .build()
+
+            return Retrofit.Builder()
+                .client(client)
+                .baseUrl(BuildConfig.BASE_URL)
+                .addConverterFactory(json.asConverterFactory("application/json".toMediaType()))
+                .build()
+                .create(MoviesApi::class.java)
+        }
+    }
+}
