--- app/src/main/java/ru/sandbox/androidacademyapp/data/ActorResults.kt
@@ -1,10 +1,0 @@
-package ru.sandbox.androidacademyapp.data
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ActorResults(
-    @SerialName("cast")
-    val actors: List<Actor>
-)
