--- app/src/main/java/ru/sandbox/androidacademyapp/data/MovieResults.kt
@@ -1,10 +1,0 @@
-package ru.sandbox.androidacademyapp.data
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class MovieResults(
-    @SerialName("results")
-    val movies: List<Movie>
-)
